requirements.txt
python-telegram-bot==20.3
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
flask==3.0.0
cryptography==43.0.0
# Calendar Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
# Email Notifications
smtplib2==0.2.1
email-validator==2.1.0
jinja2==3.1.2
# Analytics
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.26.2
# Multi-language Support
babel==2.13.1
gettext==4.0
# Additional utilities
requests==2.31.0
pytz==2023.3
dateutils==0.6.12
ical2vcs==0.3.0
